LPD-88290 Rename PanelResizer by ResizeHandle
