+++ modules/apps/frontend-js/frontend-js-clay-web/clay/clay-core/src/index.ts
@@ -12,9 +12,9 @@ export type {Item} from './language-picker';
 export type {AnnouncerAPI} from './live-announcer';
 export {Nav} from './nav';
 export {OverlayMask} from './overlay-mask';
-export {PanelResizer} from './panel-resizer';
-export type {Position} from './panel-resizer';
 export {Picker, Option} from './picker';
+export {ResizeHandle} from './resize-handle';
+export type {Position} from './resize-handle';
 export {SidePanel} from './side-panel';
 export {SidePanelWithDrilldown} from './side-panel';
 export type {SidePanelProps} from './side-panel';
