+++ modules/apps/frontend-js/frontend-js-clay-web/clay/clay-core/src/resize-handle/index.ts
@@ -3,5 +3,5 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-export {PanelResizer} from './PanelResizer';
-export type {Position} from './PanelResizer';
+export {ResizeHandle} from './ResizeHandle';
+export type {Position} from './ResizeHandle';
