+++ modules/apps/frontend-js/frontend-js-clay-web/node-scripts.config.js
@@ -80,6 +80,7 @@ module.exports = {
 			'OverlayMask',
 			'Picker',
 			'Provider',
+			'ResizeHandle',
 			'Row',
 			'SidePanel',
 			'SidePanelWithDrilldown',
